contrib/hived-config-patcher/requirements.txt
pyyaml==5.3.1
kubernetes==11.0.0
